better pixel-aligned 16px icon
Previously, the 16px icon was the vector icon (mastered at 48px) scaled down to
16px. It needed a bit more love. It was tricky to get right; ultimately used a
6px diameter lower left circle, and a 6.5px diameter upper right circle.

7px diameter circles filled the icon nicely, but such an arrangement could not
accomodate a good pixel-aligned arrow (good is when the connection between the
circle and arrow bisects the arrow's right angle). 6px for both circles yielded a
figure that was too small and cramped.

Also optimized the PNGs for size.