pull looping behavior of pingTest() out of transition()
the functionality of transition() was overloaded; it should have been only
responsible for updating the UI, but it was also the component that sent out
the next ping (and was therefore responsible for the continuous/looping nature
of pingTest()). Now, the continuous aspect of pingTest happens more explicitly
in a function called testContinuously.