Enable Refreshing for 127.0.0.1 and localhost pages (in addition to file://)
It's not uncommon for a web developer to need to view their website through a
server. I thought about switching Refreshing from a Chrome page action to a
browser action so that developers could enable Refreshing even if they weren't
using file:// URLs, but decided that enabling the page action for 127.0.0.1 and
localhost would support most use cases. In the future, I might add an options
page that lets users specify additional hostnames for which they want the
Refreshing page action enabled.